Improve stability of contrastive loss

+++ src/trainer.py
@@ -57,7 +57,7 @@ class SaintSupLightningModule(pl.LightningModule):
 
         outputs = self.fc(x[:, self.cls_token_idx, :]).squeeze()    # BS x embed_dim
         
-        # To cater for binary/multiclass training
+        # Need to cast to cater for either binary or multi-class loss
         targets = targets.float() if self.num_classes is None else targets
         
         loss = self.criterion(outputs, targets)
@@ -68,8 +68,8 @@ class SaintSupLightningModule(pl.LightningModule):
             else:
                 preds = nn.functional.softmax(outputs, dim=1)
 
-        auroc_fn.update(preds, targets)
-        accuracy_fn.update(preds, targets)
+        auroc_fn.update(preds, targets.long())
+        accuracy_fn.update(preds, targets.long())
         
         return loss
